float to float64

+++ src/portfolio.jl
@@ -21,7 +21,7 @@ end
 mutable struct Portfolio <: IS.InfrastructureSystemsType
     aggregation::Type{<:Union{PSY.ACBus, PSY.AggregationTopology}}
     discount_rate::Float64
-    inflation_rate::Float
+    inflation_rate::Float64
     base_year::Int
     data::IS.SystemData # Inputs to the model
     investment_schedule::Dict # Investment decisions container i.e., model outputs. Container TBD
